Time erledigt.k

+++ src/org/campus02/primThreads/Prim.java
@@ -14,7 +14,7 @@ public class Prim implements Runnable{
 	@Override
 	public void run() {
 		long d;
-		for (int i = 2; i < c/2; i++) {
+		for (long i = 2; i < c/2; i++) {
 			d=c%i;
 			if(d==0) {
 				isPrim = false;
